+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Add aem get_development_bundles_status method
+
 ## 3.12.0 - 2021-09-29
 ### Added
 - Add bundle is_active method